June 29th, 2010
NEW - bug 292412: [LTTng] Control Flow View Zoom-In some times does not show the state 
https://bugs.eclipse.org/bugs/show_bug.cgi?id=292412

The synthetic event provider has been updated to draw the last state registered in its corresponding state provider, 
when multiple traces are involved within the same experiment, a loop is performed to trigger the drawing of the last state on every trace.

This issue is also covering some left overs of Bug 303510, where selecting next or previous event on the first event or last event would either go all the way to the birth time of the process or the end time of the trace. These border events don't have a reference to the next event since it's out of the scope of the data collected when a zoom in / out was previously selected.

The mechanism is adjusted to artificially extend these events to either half page before or after the visible time window, so moving to previous or next border events would re-select half the window before or after and therefore be able to keep the zoom level and visualise a before or after layout of information.

The above required an adjustment to the tmf time analysis widget to correct a bug, when the currently selected time is in the middle of the last event of a particular process and the next event option is selected, this was causing to re-select back to the start time of the last event, the correction causes it to select to the end of the last event as it's expected.

git-svn-id: http://dev.eclipse.org/svnroot/technology/org.eclipse.linuxtools/lttng/trunk@25139 ee007c2a-0a25-0410-9ab9-bf268980928c