Switch the inner container for paths to Vectors
That is, vectors of strings instead of slash-delimited Strings
This way, partial paths only need to be stored once in memory.

Also added genericity to the Quark Table class so it can be re-used
elsewhere easily. We also had to provide an extension
(PathConversionTable, ok not the best name ever) to define the
specific reading and writing methods.

Oh yeah, also added a first version of the querying API.