Refactoring: change most "paths" to "attributes"
It was getting confusing seeing pathAsString, pathAsVector and
pathAsPath, especially since the last one is the only one referring
to the Path class.

Enters the notion of "attribute", shamelessly taken from LTTV.
Each attribute will refer to a 'state entry you want'.
The basic type of the attribute is Vector<String>, which is how
the outside application should represent it anyway.

We can then differiate it from the attribute-as-one-string reprensentation,
the "path" (the Path class) and the integer representation from the
quark table. They all refer to the same thing anyway.